+++ x/stake/stake.go
@@ -10,13 +10,13 @@ import (
 type (
 	Keeper                = keeper.Keeper
 	Validator             = types.Validator
+	BechValidator         = types.BechValidator
 	Description           = types.Description
 	Delegation            = types.Delegation
 	UnbondingDelegation   = types.UnbondingDelegation
 	Redelegation          = types.Redelegation
 	Params                = types.Params
 	Pool                  = types.Pool
-	PoolShares            = types.PoolShares
 	MsgCreateValidator    = types.MsgCreateValidator
 	MsgEditValidator      = types.MsgEditValidator
 	MsgDelegate           = types.MsgDelegate
@@ -62,9 +62,6 @@ var (
 
 	DefaultParams       = types.DefaultParams
 	InitialPool         = types.InitialPool
-	NewUnbondedShares   = types.NewUnbondedShares
-	NewUnbondingShares  = types.NewUnbondingShares
-	NewBondedShares     = types.NewBondedShares
 	NewValidator        = types.NewValidator
 	NewDescription      = types.NewDescription
 	NewGenesisState     = types.NewGenesisState